requirements.txt
asgiref==3.6.0
Django==4.1.4
python-environ==0.4.54
sqlparse==0.4.3
tzdata==2022.7
